# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  backend_internal_domain_models.AddItemToGroupRequest:
    properties:
      attribute_id:
        type: integer
      custom_display_name:
        type: string
      icon:
        type: string
      sort_order:
        type: integer
      visibility_condition:
        items:
          type: integer
        type: array
    required:
    - attribute_id
    type: object
...
  backend_internal_domain_models.AttachGroupToCategoryRequest:
    properties:
      collapsed_by_default:
        type: boolean
      component_id:
        type: integer
      configuration:
        items:
          type: integer
        type: array
      display_mode:
        type: string
      group_id:
        type: integer
      is_active:
        type: boolean
      sort_order:
        type: integer
    required:
    - group_id
    type: object
  backend_internal_domain_models.AttributeGroup:
    properties:
      created_at:
        type: string
      description:
        type: string
      display_name:
        type: string
      icon:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      is_system:
        type: boolean
      name:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
    type: object
  backend_internal_domain_models.AttributeGroupItem:
    properties:
      attribute:
        $ref: '#/definitions/backend_internal_domain_models.CategoryAttribute'
      attribute_id:
        type: integer
      created_at:
        type: string
      custom_display_name:
        type: string
      group_id:
        type: integer
      icon:
        type: string
      id:
        type: integer
      sort_order:
        type: integer
      visibility_condition:
        items:
          type: integer
        type: array
    type: object
...
  backend_internal_domain_models.CategoryAttribute:
    properties:
      attribute_type:
        type: string
      created_at:
        type: string
      custom_component:
        type: string
      display_name:
        type: string
      icon:
        type: string
      id:
        type: integer
      is_filterable:
        type: boolean
      is_required:
        type: boolean
      is_searchable:
        type: boolean
      name:
        type: string
      option_translations:
        additionalProperties:
          additionalProperties:
            type: string
          type: object
        type: object
      options:
        items:
          type: integer
        type: array
      sort_order:
        type: integer
      translations:
        additionalProperties:
          type: string
        type: object
      validation_rules:
        items:
          type: integer
        type: array
    type: object
...
  backend_internal_domain_models.UpdateAttributeGroupRequest:
    properties:
      description:
        type: string
      display_name:
        type: string
      icon:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      sort_order:
        type: integer
    type: object
...
  backend_pkg_utils.ErrorResponseSwag:
    properties:
      error:
        example: Описание ошибки
        type: string
    type: object
  backend_pkg_utils.SuccessResponseSwag:
    properties:
      data: {}
      success:
        example: true
        type: boolean
    type: object
...
  internal_proj_marketplace_handler.AttributeGroupResponse:
    properties:
      group:
        $ref: '#/definitions/backend_internal_domain_models.AttributeGroup'
    type: object
  internal_proj_marketplace_handler.AttributeGroupWithItemsData:
    properties:
      group:
        $ref: '#/definitions/backend_internal_domain_models.AttributeGroup'
      items:
        items:
          $ref: '#/definitions/backend_internal_domain_models.AttributeGroupItem'
        type: array
    type: object
  internal_proj_marketplace_handler.AttributeGroupsResponse:
    properties:
      groups:
        items:
          $ref: '#/definitions/backend_internal_domain_models.AttributeGroup'
        type: array
    type: object
...
  internal_proj_marketplace_handler.IDMessageResponse:
    properties:
      id:
        example: 123
        type: integer
      message:
        example: marketplace.created
        type: string
    type: object
...
  internal_proj_marketplace_handler.MessageResponse:
    properties:
      message:
        example: marketplace.operationSuccess
        type: string
      success:
        example: true
        type: boolean
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@svetu.rs
    name: API Support
    url: https://svetu.rs/support
  description: API для платформы Sve Tu - маркетплейс объявлений
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://svetu.rs/terms
  title: Sve Tu Marketplace API
  version: "1.0"
paths:
...
  /api/v1/marketplace/admin/attribute-groups:
    get:
      consumes:
      - application/json
      description: Returns all attribute groups
      produces:
      - application/json
      responses:
        "200":
          description: List of attribute groups
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_marketplace_handler.AttributeGroupsResponse'
              type: object
        "500":
          description: marketplace.listGroupsError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: List attribute groups
      tags:
      - marketplace-admin-attribute-groups
    post:
      responses:
        "400":
          description: marketplace.invalidRequest or marketplace.groupNameRequired
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: marketplace.createGroupError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
  /api/v1/marketplace/admin/attribute-groups/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes an attribute group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: marketplace.groupDeleted
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_marketplace_handler.MessageResponse'
              type: object
        "400":
          description: marketplace.invalidGroupId
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: marketplace.deleteGroupError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Delete attribute group
      tags:
      - marketplace-admin-attribute-groups
    get:
      consumes:
      - application/json
      description: Returns information about a specific attribute group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Attribute group information
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_marketplace_handler.AttributeGroupResponse'
              type: object
        "400":
          description: marketplace.invalidGroupId
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "404":
          description: marketplace.groupNotFound
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Get attribute group
      tags:
      - marketplace-admin-attribute-groups
    put:
      consumes:
      - application/json
      description: Updates an existing attribute group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated group data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/backend_internal_domain_models.UpdateAttributeGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: marketplace.groupUpdated
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_marketplace_handler.MessageResponse'
              type: object
        "400":
          description: marketplace.invalidGroupId or marketplace.invalidRequest
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: marketplace.updateGroupError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Update attribute group
      tags:
      - marketplace-admin-attribute-groups
  /api/v1/marketplace/admin/attribute-groups/{id}/items:
    get:
      consumes:
      - application/json
      description: Returns attribute group with all its attributes
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Group with items
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_marketplace_handler.AttributeGroupWithItemsData'
              type: object
        "400":
          description: marketplace.invalidGroupId
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "404":
          description: marketplace.groupNotFound
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: marketplace.getGroupItemsError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Get attribute group with items
      tags:
      - marketplace-admin-attribute-groups
    post:
      consumes:
      - application/json
      description: Adds an attribute to a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Attribute data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/backend_internal_domain_models.AddItemToGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Item added successfully
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_marketplace_handler.IDMessageResponse'
              type: object
        "400":
          description: marketplace.invalidGroupId or marketplace.invalidRequest
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: marketplace.addItemError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Add attribute to group
      tags:
      - marketplace-admin-attribute-groups
  /api/v1/marketplace/admin/attribute-groups/{id}/items/{attributeId}:
    delete:
      consumes:
      - application/json
      description: Removes an attribute from a group
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: Attribute ID
        in: path
        name: attributeId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: marketplace.itemRemoved
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_marketplace_handler.MessageResponse'
              type: object
        "400":
          description: marketplace.invalidGroupId or marketplace.invalidAttributeId
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: marketplace.removeItemError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Remove attribute from group
      tags:
      - marketplace-admin-attribute-groups
  /api/v1/marketplace/admin/categories/{id}/groups:
    get:
      consumes:
      - application/json
      description: Returns attribute groups attached to a category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Category groups
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_marketplace_handler.AttributeGroupsResponse'
              type: object
        "400":
          description: marketplace.invalidCategoryId
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: marketplace.getCategoryGroupsError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Get category attribute groups
      tags:
      - marketplace-admin-attribute-groups
    post:
      consumes:
      - application/json
      description: Attaches an attribute group to a category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Group attachment data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/backend_internal_domain_models.AttachGroupToCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Group attached successfully
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_marketplace_handler.IDMessageResponse'
              type: object
        "400":
          description: marketplace.invalidCategoryId or marketplace.invalidRequest
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: marketplace.attachGroupError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Attach group to category
      tags:
      - marketplace-admin-attribute-groups
  /api/v1/marketplace/admin/categories/{id}/groups/{groupId}:
    delete:
      consumes:
      - application/json
      description: Detaches an attribute group from a category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: marketplace.groupDetached
          schema:
            allOf:
            - $ref: '#/definitions/backend_pkg_utils.SuccessResponseSwag'
            - properties:
                data:
                  $ref: '#/definitions/internal_proj_marketplace_handler.MessageResponse'
              type: object
        "400":
          description: marketplace.invalidCategoryId or marketplace.invalidGroupId
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
        "500":
          description: marketplace.detachGroupError
          schema:
            $ref: '#/definitions/backend_pkg_utils.ErrorResponseSwag'
      security:
      - BearerAuth: []
      summary: Detach group from category
      tags:
      - marketplace-admin-attribute-groups
...
securityDefinitions:
  Bearer:
    description: Bearer token для авторизации
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"